backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
sqlalchemy[asyncio]>=2.0.30
aiosqlite>=0.20.0
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.7.1
pydantic-settings>=2.3.0
email-validator>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
argon2-cffi>=23.1.0
python-multipart>=0.0.9
redis[asyncio]>=5.0.4
aiofiles>=23.2.1
Pillow>=11.0.0
reportlab>=4.1.0
httpx>=0.27.0
slowapi>=0.1.9
structlog>=24.2.0
pytest>=8.2.0
pytest-asyncio>=0.23.7
pytest-httpx>=0.30.0
